[Arm64/Linux] PAL_GetJitCpuCapabilityFlags()
Initial PAL API definition
Implementation for ARM64/Linux
