+++ src/pal/src/config.h.in
@@ -23,6 +23,7 @@
 #cmakedefine01 HAVE_PRCTL_H
 #cmakedefine01 HAVE_NUMA_H
 #cmakedefine01 HAVE_PTHREAD_NP_H
+#cmakedefine01 HAVE_AUXV_HWCAP_H
 
 #cmakedefine01 HAVE_KQUEUE
 #cmakedefine01 HAVE_PTHREAD_SUSPEND
